+++ tests/test_feed.py
@@ -22,6 +22,88 @@ class TestFeed(unittest.TestCase):
             else:
                 self.assertIsNone(value)
 
+    def test_eq(self):  
+        self.assertEqual(Feed(), Feed())
+
+        feed1 = Feed(stops=pd.DataFrame([[1, 2], [3, 4]], columns=['a', 'b']))
+        self.assertEqual(feed1, feed1)
+
+        feed2 = Feed(stops=pd.DataFrame([[4, 3], [2, 1]], columns=['b', 'a']))
+        self.assertEqual(feed1, feed2)
+        
+        feed2 = Feed(stops=pd.DataFrame([[3, 4], [2, 1]], columns=['b', 'a']))
+        self.assertNotEqual(feed1, feed2)
+    
+        feed2 = Feed(stops=pd.DataFrame([[4, 3], [2, 1]], columns=['b', 'a']),
+          dist_units_in='km')
+        self.assertEqual(feed1, feed2)
+
+        feed2 = Feed(stops=pd.DataFrame([[4, 3], [2, 1]], columns=['b', 'a']),
+          dist_units_in='mi')
+        self.assertNotEqual(feed1, feed2)
+
+    # --------------------------------------------
+    # Test functions about basics
+    # --------------------------------------------
+    def test_copy(self):
+        feed1 = read_gtfs('data/cairns_gtfs.zip')
+        feed2 = copy(feed1)
+        for key, value in feed2.__dict__.items():
+            expect_value = getattr(feed1, key)            
+            if isinstance(value, pd.DataFrame):
+                assert_frame_equal(value, expect_value)
+            elif isinstance(value, pd.core.groupby.DataFrameGroupBy):
+                self.assertEqual(value.groups, expect_value.groups)
+            elif isinstance(value, FunctionType):
+                # No need to check this
+                continue
+            else:
+                self.assertEqual(value, expect_value)
+
+    # def test_concatenate(self):
+    #     import re 
+
+
+    #     # Trivial cases
+    #     cat = concatenate([])
+    #     self.assertEqual(cat, Feed())
+
+    #     feed = read_gtfs('data/cairns_gtfs.zip')
+    #     cat = concatenate([feed])
+    #     self.assertEqual(cat, feed)
+        
+    #     # Nontrivial cases
+    #     n = 3
+    #     feeds = [feed for i in range(n)]
+    #     prefixes = ['{!s}_'.format(i) for i in range(n)]
+
+    #     def strip_prefix(x):
+    #         if pd.notnull(x):
+    #             x = re.sub(r'^\d\_', '', x)
+    #         return x
+
+    #     cat = concatenate(feeds, prefixes)
+    #     for key in cs.FEED_INPUTS:
+    #         x = getattr(feed, key)
+    #         y = getattr(cat, key)
+    #         if isinstance(x, pd.DataFrame):
+    #             # Data frames should be the correct shape
+    #             self.assertEqual(y.shape[0], n*x.shape[0])
+    #             self.assertEqual(y.shape[1], x.shape[1])
+    #             # Stripping prefixes should yield original data frame
+    #             print(y.dtypes)
+    #             for col in cs.ID_COLUMNS:
+    #                 if col in y.columns:    
+    #                     y[col] = y[col].map(strip_prefix)
+    #             y = y.drop_duplicates()
+    #             print(ut.almost_equal(y, x), y.equals(x))
+    #             print(y.dtypes)
+    #             print(x.dtypes)
+    #             self.assertTrue(ut.almost_equal(y, x))
+    #         else:
+    #             # Non data frame values should be equal
+    #             self.assertEqual(x, y)
+
     # --------------------------------------------
     # Test functions about inputs and outputs
     # --------------------------------------------
@@ -79,20 +161,3 @@ class TestFeed(unittest.TestCase):
         # Remove extracted directory
         shutil.rmtree(dir_name)
 
-    # --------------------------------------------
-    # Test functions about basics
-    # --------------------------------------------
-    def test_copy(self):
-        feed1 = read_gtfs('data/cairns_gtfs.zip')
-        feed2 = copy(feed1)
-        for key, value in feed2.__dict__.items():
-            expect_value = getattr(feed1, key)            
-            if isinstance(value, pd.DataFrame):
-                assert_frame_equal(value, expect_value)
-            elif isinstance(value, pd.core.groupby.DataFrameGroupBy):
-                self.assertEqual(value.groups, expect_value.groups)
-            elif isinstance(value, FunctionType):
-                # No need to check this
-                continue
-            else:
-                self.assertEqual(value, expect_value)