+++ tests/test_utilities.py
@@ -91,6 +91,16 @@ class TestUtilities(unittest.TestCase):
         expect = [0, 1]
         assert_array_equal(get, expect)
 
+    def test_almost_equal(self):
+        f = pd.DataFrame([[1, 2], [3, 4]], columns=['a', 'b'])
+        self.assertTrue(almost_equal(f, f))
+        g = pd.DataFrame([[4, 3], [2, 1]], columns=['b', 'a'])
+        self.assertTrue(almost_equal(f, g))
+        h = pd.DataFrame([[1, 2], [5, 4]], columns=['a', 'b'])
+        self.assertFalse(almost_equal(f, h))
+        h = pd.DataFrame()
+        self.assertFalse(almost_equal(f, h))
+
 
 if __name__ == '__main__':
     unittest.main()